pyproject.toml
[project]
name = "sata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain-groq>=1.1.2",
    "langchian>=0.2.5",
    "langgraph>=1.0.8",
    "langgraph-cli[inmem]>=0.4.12",
    "pandas>=2.3.3",
    "pysqlite3>=0.6.0",
    "python-dotenv>=1.2.1",
]
